+++ egui_demo_lib/src/apps/demo/tests.rs
@@ -179,6 +179,7 @@ pub struct TableTest {
     num_rows: usize,
     min_col_width: f32,
     max_col_width: f32,
+    text_length: usize,
 }
 
 impl Default for TableTest {
@@ -188,6 +189,7 @@ impl Default for TableTest {
             num_rows: 4,
             min_col_width: 10.0,
             max_col_width: 200.0,
+            text_length: 10,
         }
     }
 }
@@ -246,6 +248,48 @@ impl super::View for TableTest {
                     ui.end_row();
                 }
             });
+
+        ui.separator();
+        ui.add(egui::Slider::new(&mut self.text_length, 1..=40).text("Text length"));
+        egui::Grid::new("parent grid").striped(true).show(ui, |ui| {
+            ui.vertical(|ui| {
+                ui.label("Vertical nest1");
+                ui.label("Vertical nest2");
+            });
+            ui.label("First row, second column");
+            ui.end_row();
+
+            ui.horizontal(|ui| {
+                ui.label("Horizontal nest1");
+                ui.label("Horizontal nest2");
+            });
+            ui.label("Second row, second column");
+            ui.end_row();
+
+            ui.scope(|ui| {
+                ui.label("Scope nest 1");
+                ui.label("Scope nest 2");
+            });
+            ui.label("Third row, second column");
+            ui.end_row();
+
+            egui::Grid::new("nested grid").show(ui, |ui| {
+                ui.label("Grid nest11");
+                ui.label("Grid nest12");
+                ui.end_row();
+                ui.label("Grid nest21");
+                ui.label("Grid nest22");
+                ui.end_row();
+            });
+            ui.label("Fourth row, second column");
+            ui.end_row();
+
+            let mut dyn_text = String::from("O");
+            dyn_text.extend(std::iter::repeat('h').take(self.text_length));
+            ui.label(dyn_text);
+            ui.label("Fifth row, second column");
+            ui.end_row();
+        });
 
         ui.vertical_centered(|ui| {
             egui::reset_button(ui, self);